Canviats tots els atributs id de int a long

+++ src/model/Candidat.java
@@ -1,14 +1,14 @@
 package model;
 
 public class Candidat {
-    int candidat_id;
-    int candidatura_id;
-    int persona_id;
-    int provincia_id;
+    long candidat_id;
+    long candidatura_id;
+    long persona_id;
+    long provincia_id;
     int num_ordre;
     char tipus;
 
-    public Candidat(int candidat_id, int candidatura_id, int persona_id, int provincia_id, int num_ordre, char tipus) {
+    public Candidat(long candidat_id, long candidatura_id, long persona_id, long provincia_id, int num_ordre, char tipus) {
         this.candidat_id = candidat_id;
         this.candidatura_id = candidatura_id;
         this.persona_id = persona_id;
@@ -17,35 +17,35 @@ public class Candidat {
         this.tipus = tipus;
     }
 
-    public int getCandidat_id() {
+    public long getCandidat_id() {
         return candidat_id;
     }
 
-    public void setCandidat_id(int candidat_id) {
+    public void setCandidat_id(long candidat_id) {
         this.candidat_id = candidat_id;
     }
 
-    public int getCandidatura_id() {
+    public long getCandidatura_id() {
         return candidatura_id;
     }
 
-    public void setCandidatura_id(int candidatura_id) {
+    public void setCandidatura_id(long candidatura_id) {
         this.candidatura_id = candidatura_id;
     }
 
-    public int getPersona_id() {
+    public long getPersona_id() {
         return persona_id;
     }
 
-    public void setPersona_id(int persona_id) {
+    public void setPersona_id(long persona_id) {
         this.persona_id = persona_id;
     }
 
-    public int getProvincia_id() {
+    public long getProvincia_id() {
         return provincia_id;
     }
 
-    public void setProvincia_id(int provincia_id) {
+    public void setProvincia_id(long provincia_id) {
         this.provincia_id = provincia_id;
     }
